+++ src/main/java/org/woehlke/bloodmoney/application/test/BloodMoneyTestService.java
@@ -1,0 +1,6 @@
+package org.woehlke.bloodmoney.application.test;
+
+public interface BloodMoneyTestService {
+
+    void createTestData();
+}