+++ src/main/java/org/woehlke/bloodmoney/measurements/BloodPressureMeasurementService.java
@@ -1,31 +1,10 @@
 package org.woehlke.bloodmoney.measurements;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import org.woehlke.bloodmoney.common.BloodMoneyEntityService;
 
-import java.util.List;
-import java.util.Optional;
+public interface BloodPressureMeasurementService extends BloodMoneyEntityService<BloodPressureMeasurementEntity,Long> {
 
-public interface BloodPressureMeasurementService {
-
-    Page<BloodPressureMeasurement> getAll(Pageable pageable);
-
-    BloodPressureMeasurement add(BloodPressureMeasurement one);
-
-    BloodPressureMeasurement update(BloodPressureMeasurement one);
-
-    BloodPressureMeasurement update(BloodPressureMeasurement one, long id);
-
-    void delete(BloodPressureMeasurement one);
-
-    List<BloodPressureMeasurement> getAll();
-
-    BloodPressureMeasurement getOne(long id);
-
-    long count();
-
-    void deleteAll();
-
-    Optional<BloodPressureMeasurement> findById(Long id);
+    @Deprecated
+    BloodPressureMeasurementEntity update(BloodPressureMeasurementEntity one);
 
 }