+++ src/main/java/org/woehlke/bloodmoney/test/TestService.java
@@ -1,6 +1,0 @@
-package org.woehlke.bloodmoney.test;
-
-public interface TestService {
-
-    void createTestData();
-}