+++ src/main/java/org/woehlke/bloodmoney/user/BloodMoneyUserAccountAuthorizationService.java
@@ -1,0 +1,8 @@
+package org.woehlke.bloodmoney.user;
+
+public interface BloodMoneyUserAccountAuthorizationService {
+
+    boolean confirmUserByLoginAndPassword(String userEmail, String userPassword);
+
+    boolean authorize(LoginFormBean loginFormBean);
+}