+++ src/main/java/org/woehlke/bloodmoney/user/UserSessionService.java
@@ -1,0 +1,7 @@
+package org.woehlke.bloodmoney.user;
+
+import org.springframework.ui.Model;
+
+public interface UserSessionService {
+    Model handleUserSession(UserSessionBean userSessionBean, Model model);
+}