+++ src/main/java/org/woehlke/bloodmoney/user/account/UserAccountSecurityService.java
@@ -1,6 +1,0 @@
-package org.woehlke.bloodmoney.user.account;
-
-import org.springframework.security.core.userdetails.UserDetailsService;
-
-public interface UserAccountSecurityService  extends UserDetailsService {
-}