+++ src/main/java/org/woehlke/bloodmoney/user/session/UserSessionService.java
@@ -1,7 +1,0 @@
-package org.woehlke.bloodmoney.user.session;
-
-import org.springframework.ui.Model;
-
-public interface UserSessionService {
-    Model handleUserSession(UserSession userSession, Model model);
-}