+++ src/test/java/org/woehlke/bloodmoney/config/BloodMoneyPropertiesTest.java
@@ -1,0 +1,4 @@
+package org.woehlke.bloodmoney.config;
+
+public class BloodMoneyPropertiesTest {
+}