+++ src/test/java/org/woehlke/bloodmoney/measurements/WebMockTest.java
@@ -1,11 +1,7 @@
 package org.woehlke.bloodmoney.measurements;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import org.woehlke.bloodmoney.user.session.UserSessionService;
+import org.woehlke.bloodmoney.user.UserSessionService;
 
 import java.util.ArrayList;
 
@@ -30,7 +26,7 @@ public class WebMockTest {
 
     //@Test
     public void greetingShouldReturnMessageFromService() throws Exception {
-        when(bloodPressureMeasurementService.getAll()).thenReturn(new ArrayList<BloodPressureMeasurement>());
+        when(bloodPressureMeasurementService.getAll()).thenReturn(new ArrayList<BloodPressureMeasurementEntity>());
         this.mockMvc.perform(get("/greeting")).andDo(print()).andExpect(status().isOk())
             .andExpect(content().string(containsString("Hello, Mock")));
     }