+++ src/test/java/org/woehlke/bloodmoney/user/UserAccountResourceTest.java
@@ -1,4 +1,0 @@
-package org.woehlke.bloodmoney.user;
-
-public class UserAccountResourceTest {
-}